+++ examples/plot_fizzbuzz.py
@@ -1,0 +1,16 @@
+"""
+Play FizzBuzz
+=============
+
+First introduced by Reginald Braithwaite and made famous by the Coding Horror
+blog.
+
+Author: Marijn van Vliet <w.m.vanvliet@gmail.com>
+"""
+
+# First we import the required modules.
+from kirpputori import fizzbuzz
+
+###############################################################################
+# Now we can play FizzBuzz for a long time!
+fizzbuzz(10_000)